// packages/js/data/src/payment-settings/action-types.ts
export enum ACTION_TYPES {
	GET_PAYMENT_GATEWAY_SUGGESTIONS_REQUEST = 'GET_PAYMENT_GATEWAY_SUGGESTIONS_REQUEST',
	GET_PAYMENT_GATEWAY_SUGGESTIONS_SUCCESS = 'GET_PAYMENT_GATEWAY_SUGGESTIONS_SUCCESS',
	GET_PAYMENT_GATEWAY_SUGGESTIONS_ERROR = 'GET_PAYMENT_GATEWAY_SUGGESTIONS_ERROR',
	ENABLE_PAYMENT_GATEWAY_REQUEST = 'ENABLE_PAYMENT_GATEWAY_REQUEST',
	ENABLE_PAYMENT_GATEWAY_SUCCESS = 'ENABLE_PAYMENT_GATEWAY_SUCCESS',
	ENABLE_PAYMENT_GATEWAY_ERROR = 'ENABLE_PAYMENT_GATEWAY_ERROR',
}